requirements.txt
fastapi
sqlalchemy
python-jose
pydantic
uvicorn
passlib
starlette
geopy
requests
psycopg2-binary
bcrypt